feat: add support to login to apple
